move funcitonality to Variables
