+++ _unittest/test_09_VariableManager.py
@@ -6,7 +6,7 @@ from _unittest.conftest import desktop_version
 import pytest
 
 from pyaedt import MaxwellCircuit
-from pyaedt.application.Variables import Variable
+from pyaedt.application.Variables import Variable, generate_validation_errors, get_value_from_string
 from pyaedt.application.Variables import decompose_variable_value
 from pyaedt.generic.general_methods import isclose
 from pyaedt.modeler.geometry_operators import GeometryOperators
@@ -16,6 +16,48 @@ from pyaedt.modeler.geometry_operators import GeometryOperators
 def aedtapp(add_app):
     app = add_app(project_name="Test_09")
     return app
+
+
+@pytest.fixture()
+def validation_input():
+    property_names = [
+        "+X Padding Type",
+        "+X Padding Data",
+        "-X Padding Type",
+        "-X Padding Data",
+        "+Y Padding Type",
+        "+Y Padding Data",
+        "-Y Padding Type",
+        "-Y Padding Data",
+        "+Z Padding Type",
+        "+Z Padding Data",
+        "-Z Padding Type",
+        "-Z Padding Data",
+    ]
+    expected_settings = [
+        "Absolute Offset",
+        "10mm",
+        "Percentage Offset",
+        "100",
+        "Transverse Percentage Offset",
+        "100",
+        "Percentage Offset",
+        "10",
+        "Absolute Offset",
+        "50mm",
+        "Absolute Position",
+        "-600mm",
+    ]
+    actual_settings = list(expected_settings)
+    return property_names, expected_settings, actual_settings
+
+
+@pytest.fixture()
+def validation_float_input():
+    property_names = ["+X Padding Data", "-X Padding Data", "+Y Padding Data"]
+    expected_settings = [100, 200.1, 300]
+    actual_settings = list(expected_settings)
+    return property_names, expected_settings, actual_settings
 
 
 class TestClass:
@@ -510,3 +552,116 @@ class TestClass:
         assert self.aedtapp.variable_manager["v2"].decompose() == (6.0, "mm")
         assert self.aedtapp.variable_manager.decompose("5mm") == (5.0, "mm")
         assert self.aedtapp.number_with_units(3.0, "mil") == "3.0mil"
+
+    def test_21_test_validator_exact_match(self, validation_input):
+        property_names, expected_settings, actual_settings = validation_input
+        validation_errors = generate_validation_errors(property_names, expected_settings, actual_settings)
+        assert len(validation_errors) == 0
+
+    def test_22_test_validator_tolerance(self, validation_input):
+        property_names, expected_settings, actual_settings = validation_input
+
+        # Small difference should produce no validation errors
+        actual_settings[1] = "10.0000000001mm"
+        actual_settings[3] = "100.0000000001"
+        actual_settings[5] = "100.0000000001"
+        validation_errors = generate_validation_errors(property_names, expected_settings, actual_settings)
+
+        assert len(validation_errors) == 0
+
+    def test_23_test_validator_invalidate_offset_type(self, validation_input):
+        property_names, expected_settings, actual_settings = validation_input
+
+        # Are expected to be "Absolute Offset"
+        actual_settings[0] = "Percentage Offset"
+
+        validation_errors = generate_validation_errors(property_names, expected_settings, actual_settings)
+
+        assert len(validation_errors) == 1
+
+    def test_24_test_validator_invalidate_offset_value(self, validation_input):
+        property_names, expected_settings, actual_settings = validation_input
+
+        # Above tolerance
+        actual_settings[1] = "10.000002mm"
+
+        validation_errors = generate_validation_errors(property_names, expected_settings, actual_settings)
+
+        assert len(validation_errors) == 1
+
+    def test_25_test_validator_invalidate_multiple(self, validation_input):
+        property_names, expected_settings, actual_settings = validation_input
+
+        actual_settings[0] = "Percentage Offset"
+        actual_settings[1] = "22mm"
+        actual_settings[2] = "Transverse Percentage Offset"
+
+        validation_errors = generate_validation_errors(property_names, expected_settings, actual_settings)
+
+        assert len(validation_errors) == 3
+
+    def test_26_test_validator_invalidate_wrong_type(self, validation_input):
+        property_names, expected_settings, actual_settings = validation_input
+
+        actual_settings[1] = "nonnumeric"
+
+        validation_errors = generate_validation_errors(property_names, expected_settings, actual_settings)
+
+        assert len(validation_errors) == 1
+
+    def test_27_test_validator_float_type(self, validation_float_input):
+        property_names, expected_settings, actual_settings = validation_float_input
+
+        validation_errors = generate_validation_errors(property_names, expected_settings, actual_settings)
+
+        assert len(validation_errors) == 0
+
+    def test_28_test_validator_float_type_tolerance(self, validation_float_input):
+        property_names, expected_settings, actual_settings = validation_float_input
+
+        # Set just below the tolerance to pass the check
+        actual_settings[0] *= 1 + 0.99 * 1e-9
+        actual_settings[1] *= 1 - 0.99 * 1e-9
+        actual_settings[2] *= 1 + 0.99 * 1e-9
+
+        validation_errors = generate_validation_errors(property_names, expected_settings, actual_settings)
+
+        assert len(validation_errors) == 0
+
+    def test_29_test_validator_float_type_invalidate(self, validation_float_input):
+        property_names, expected_settings, actual_settings = validation_float_input
+
+        # Set just above the tolerance to fail the check
+        actual_settings[0] *= 1 + 1.01 * 1e-9
+        actual_settings[1] *= 1 + 1.01 * 1e-9
+        actual_settings[2] *= 1 + 1.01 * 1e-9
+
+        validation_errors = generate_validation_errors(property_names, expected_settings, actual_settings)
+
+        assert len(validation_errors) == 3
+
+    def test_30_test_validator_float_type_invalidate(self, validation_float_input):
+        property_names, expected_settings, actual_settings = validation_float_input
+
+        actual_settings[0] *= 2
+
+        validation_errors = generate_validation_errors(property_names, expected_settings, actual_settings)
+
+        assert len(validation_errors) == 1
+
+    def test_31_test_checknumber_valid_input(self):
+        is_number, value = get_value_from_string("10.1mm")
+
+        assert is_number
+        assert value == 10.1
+
+    def test_32_test_checknumber_valid_input(self):
+        is_number, value = get_value_from_string("10mm")
+
+        assert is_number
+        assert value == 10.0
+
+    def test_33_test_checknumber_invalid_input(self):
+        is_number, _ = get_value_from_string("test_mm")
+
+        assert not is_number
