+++ pyaedt/application/Variables.py
@@ -15,6 +15,7 @@ Examples
 
 from __future__ import absolute_import  # noreorder
 from __future__ import division
+import math
 
 import os
 import re
@@ -251,6 +252,28 @@ def _find_units_in_dependent_variables(variable_value, full_variables={}):
 
 
 @pyaedt_function_handler()
+def get_value_from_string(input_string : str):
+    """Return a tuple with a Boolean indicating whether the given string contains a number, and
+    a float value with the value. If the input string does not contains a string the value will be 0.
+
+    Parameters
+    ----------
+    input_string : str
+
+    Returns
+    -------
+    tuples
+        A tuple with a Boolean indicating whether the given string contains a number, 
+        and the float value of the input string, if available
+    """
+    match = re.search(r"[-+]?(?:\d*\.*\d+)", input_string)
+    if match:
+        return True, float(match.group())
+    else:
+        return False, 0.0
+
+
+@pyaedt_function_handler()
 def decompose_variable_value(variable_value, full_variables={}):
     """Decompose a variable value.
 
@@ -291,6 +314,36 @@ def decompose_variable_value(variable_value, full_variables={}):
                     float_value = variable_value
 
     return float_value, units
+
+
+@pyaedt_function_handler()
+def generate_property_validation_errors(property_name, expected, actual):
+    if isinstance(expected, (float, int)) and isinstance(actual, (float, int)):
+        if not math.isclose(expected, actual):
+            yield "Error {0}: Expected {1}, got {2}".format(property_name, expected, actual)
+    elif isinstance(expected, str) and isinstance(actual, str):
+        expected_is_numeric, expected_value = get_value_from_string(expected)
+        actual_is_numeric, actual_value = get_value_from_string(actual)
+        if expected_is_numeric != actual_is_numeric:
+            yield "Error {0}: Cannot match '{1}', with '{2}'".format(property_name, expected, actual)
+        elif expected_is_numeric:
+            if not math.isclose(expected_value, actual_value):
+                yield "Error {0}: Expected {1}, got {2}".format(property_name, expected, actual)
+        else:
+            if expected != actual:
+                yield "Error {0}: Expected {1}, got {2}".format(property_name, expected, actual)
+    else:
+        if expected != actual:
+            yield "Error {0}: Expected {1}, got {2}".format(property_name, expected, actual)
+
+
+def generate_validation_errors(property_names, expected_settings, actual_settings):
+    validation_errors = [
+        error
+        for property_name, expected, actual in zip(property_names, expected_settings, actual_settings)
+        for error in generate_property_validation_errors(property_name, expected, actual)
+    ]
+    return validation_errors
 
 
 class VariableManager(object):
