+++ pyaedt/modeler/modeler3d.py
@@ -3,10 +3,9 @@ from __future__ import absolute_import  # noreorder
 import copy
 import datetime
 import json
-import math
 import os.path
-import re
 import warnings
+from pyaedt.application.Variables import generate_validation_errors
 
 from pyaedt.generic.general_methods import GrpcApiError
 from pyaedt.generic.general_methods import generate_unique_name
@@ -1366,44 +1365,6 @@ class Modeler3D(GeometryModeler, Primitives3D, object):
         else:
             return segment_objects
 
-    def _check_value_from_string(input_string):
-        """Return a tuple with a Boolean indicating whether the given string contains a number, and
-        a float value with the value. If the input string does not contains a string the value will be 0.
-        """
-        match = re.search(r"[-+]?(?:\d*\.*\d+)", input_string)
-        if match:
-            return True, float(match.group())
-        else:
-            return False, 0.0
-
-    def _generate_property_validation_errors(property_name, expected, actual):
-        if isinstance(expected, (float, int)) and isinstance(actual, (float, int)):
-            if not math.isclose(expected, actual):
-                yield "Error {0}: Expected {1}, got {2}".format(property_name, expected, actual)
-        elif isinstance(expected, str) and isinstance(actual, str):
-            expected_is_numeric, expected_value = Modeler3D._check_value_from_string(expected)
-            actual_is_numeric, actual_value = Modeler3D._check_value_from_string(actual)
-
-            if expected_is_numeric != actual_is_numeric:
-                yield "Error {0}: Cannot match '{1}', with '{2}'".format(property_name, expected, actual)
-            elif expected_is_numeric:
-                if not math.isclose(expected_value, actual_value):
-                    yield "Error {0}: Expected {1}, got {2}".format(property_name, expected, actual)
-            else:
-                if expected != actual:
-                    yield "Error {0}: Expected {1}, got {2}".format(property_name, expected, actual)
-        else:
-            if expected != actual:
-                yield "Error {0}: Expected {1}, got {2}".format(property_name, expected, actual)
-
-    def _generate_validation_errors(property_names, expected_settings, actual_settings):
-        validation_errors = [
-            error
-            for property_name, expected, actual in zip(property_names, expected_settings, actual_settings)
-            for error in Modeler3D._generate_property_validation_errors(property_name, expected, actual)
-        ]
-
-        return validation_errors
 
     @pyaedt_function_handler
     def change_region_padding(self, padding_data, padding_type, direction=None, region_name="Region"):
@@ -1487,7 +1448,7 @@ class Modeler3D(GeometryModeler, Primitives3D, object):
             property_names = [lst[0].strip("NAME:") for lst in modify_props]
             actual_settings = [create_region.GetPropValue(property_name) for property_name in property_names]
             expected_settings = [lst[-1] for lst in modify_props]
-            validation_errors = Modeler3D._generate_validation_errors(
+            validation_errors = generate_validation_errors(
                 property_names, expected_settings, actual_settings
             )
